# Increase performance of nested keys search