This follows PR #253, which closed Issue #249 induced by PR #195.

It generalizes to nested keys the improvement from PR #253 which consists in assuming that all keys are lower-cased, thus allowing to search for values by writing
```golang
val = m[key]
```
instead of
```golang
for k, v := range m {
  if strings.ToLower(k) == strings.ToLower(key) {
    val = v
  }
}
```
This induces a huge performance gain.

Function `insensitiviseMap()` was also fixed (it did not recurse correctly through the inner maps).

@bep my performance tests show no performance loss (and even very little gain) from your version, can I let you benchmark on your hardware to make sure I didn't remove anything essential (it should be fine, but let's double-check this time...)?

If that's OK for you, I think this would be a more generic solution than your patch, less likely to be bypassed in the future.