@bep 2016-10-12
This looks very good in general, but it is slightly slower...

```
▶ benchcmp /tmp/1.bench /tmp/2.bench
benchmark                     old ns/op     new ns/op     delta
BenchmarkGetBool-4            553           562           +1.63%
BenchmarkGet-4                452           477           +5.53%
BenchmarkGetBoolFromMap-4     7.00          6.51          -7.00%

benchmark                     old allocs     new allocs     delta
BenchmarkGetBool-4            4              5              +25.00%
BenchmarkGet-4                4              5              +25.00%
BenchmarkGetBoolFromMap-4     0              0              +0.00%

benchmark                     old bytes     new bytes     delta
BenchmarkGetBool-4            49            65            +32.65%
BenchmarkGet-4                48            64            +33.33%
BenchmarkGetBoolFromMap-4     0             0             +0.00%
```

So, if you reintroduce my `searchMapForKey` method the above be fine.